- Handle delay inter after watching rewarded ads

+++ Assets/_Modules/_Game/Scripts/BaseGameContext.cs
@@ -709,7 +709,7 @@ namespace Mimi.Prototypes
                 .SetInt(ConfigKey.HardLevelBaseTime, 30)
                 .SetInt(ConfigKey.HardLevelAdditionalTime, 30)
                 .SetInt(ConfigKey.HardLevelWarningTime, 10)
-                .SetInt(ConfigKey.CooldownInterAfterShowReward, 30)
+                .SetFloat(ConfigKey.CooldownInterAfterShowReward, 30)
                 .SetInt(ConfigKey.LifeCooldown, 900)
                 .SetInt(ConfigKey.MaxLife, 10)
                 .SetInt(ConfigKey.LifeRecoverAfterChapter, 2)
